Allow creating new Builders using BaseRecordBuilder
The current `BaseRecordBuilder` interface defines some abstract methods,
without declaring their visibility.

This means it is by default, they are `package-private`, and we can't
create sub-classes extending the behaviour.

One scenario where extanding from another package is ideal would be when
extra headers are required to be sent to Kafka. This commit changes the
modifier to be public so others could extend them.