fix: fix tests
There were two issues preventing the tests from running:

1) The tests use the 'latest' image tag for MQ - and watch for a
particular message code in the log to recognise when the queue
manager is ready. Changes in a recent version of MQ meant that
the log output has changed, so the tests were waiting forever for
a log message that is no longer output.
I've updated the message code to match the current MQ behaviour.

2) Fix for a regression that caused a NullPointerException when
the MQSourceTaskIT test cleanup method ran - by renaming the
locally-scoped SourceTask variable to use the class-scoped
variable that the cleanup depends on.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>