fix: fix the no-auth MQ integration tests
We use the latest tag MQ image in our integration tests. In
December, MQ's container image removed the no-auth svrconn channel
so this breaks the connector tests that try to make connections
to MQ without credentials.

This commit adds a custom MQSC script to configure the queue
manager to restore the previous behaviour.

The MQSourceTaskAuthIT tests still test the ability to connect to
a queue manager with auth credentials, so this commit means we
still test connecting with and without credentials.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>