fix: use of config in tests
Some of the existing tests create a JsonRecordBuilder and start
using it without configuring it. This is not valid and would not
happen in real use as configure() is a required step in the
Connect lifecycle.

As I want to introduce new config for the record builder, I need
to correct this mistake first.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>