feat: support for schemas in JSON record builder
This commit introduces support for emitting structured records
from the JSON record builder. This will allow the MQ Source
Connector to read JSON string messages from MQ, and produce them
to Kafka using any standard Converter (e.g. to produce them in
Avro or Protobuf formats if desired).

The JsonConverter dependency used in JSON record builder has
support for this from Kafka Connect v4.2, so the simplest
implementation would be to update the dependency in pom.xml
to version 4.2, and just pass through the schemas.enable and
schema.content configuration properties to the converter
and leave the Converter to do everything.

This felt like an overly aggressive dependency jump, so in the
interest of continuing to support Connect 3.x versions, I've
implemented a fall-back implementation that reuses the schema
"envelope" approach present in JsonConverter 3.x
The additional string operations this will incur for every
message will almost certainly impact performance, so I see this
as a temporary workaround that we should remove as soon as we
feel that Connect 4.x adoption is sufficient.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>