chore: revert to Java 8 syntax
I didn't realise we were still using Java 8 to build the
connector. Given that it has been deprecated in Kafka since v3.0
this doesn't seem like the right choice, but that feels like too
significant a change to include in this pull request, so for now
I'll just remove the newer syntax I'd used.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>