feat: support for schemas in JSON record builder (#174)
This pull request introduces support for emitting structured records from 
the JSON record builder. This will allow the MQ Source Connector to read 
JSON string messages from MQ, and produce them to Kafka using any standard 
Converter (e.g. to produce them in Avro or Protobuf formats if desired).

I've chosen to support Kafka Connect's JSON schema support, rather than 
the different (and more widely understood) JSON schema.

While supporting "standard" JSON schema would have simplified the user 
config in some respects, this would have left the MQ Connector with 
responsibility of performing the (ambiguous) conversion from the 
user-provided JSON schema to the schema used in Connect. As there is not 
a 1:1 mapping between these two schema types, I think it would be 
difficult to do such a conversion in a way that always meets user 
expectations.

Instead, by making the user provide a Connect JSON schema, I'm proposing 
forcing the user to manually convert any json schema they may already 
have into a Connect schema - forcing them to make the appropriate 
choices in mapping between the two type systems.

This was a difficult trade-off to make, as I'm favouring unambiguity of 
config over ease of config (if we assume that more users are comfortable 
writing "standard" JSON schemas than Connect JSON schemas). To try and 
catch confusions in this, I've included validation to ensure that we 
reject non-Connect schemas.

The JsonConverter dependency used in JSON record builder has support for 
this from Kafka Connect v4.2, so the simplest implementation would be to 
update the dependency in pom.xml to version 4.2, and just pass through 
the schemas.enable and schema.content configuration properties to the 
converter and leave the Converter to do everything.

This felt like an overly aggressive dependency jump, so in the interest 
of continuing to support Connect 3.x versions, I've implemented a 
fall-back implementation that reuses the schema "envelope" approach 
present in JsonConverter 3.x 

The additional string operations this will incur for every message will 
almost certainly impact performance, so I see this as a temporary 
workaround that we should remove as soon as we feel that Connect 4.x 
adoption is sufficient.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>